Update Appointment screen and add TopDoctorAppointment and AllTopDoctorAppointment screens
The Appointment screen has been updated and two new screens, TopDoctorAppointment and AllTopDoctorAppointment, have been added. These new screens are used to display the appointment details for a specific top doctor and all top doctors, respectively. The new screens are used in the TopDoctorDetail and AllTopDoctorDetail components, and can be accessed from the DoctorDetail and AllDoctorDetail screens. The Appointment screen is no longer used and has been replaced with the new screens.